Moved metadata module under index
