+++ src/index/metadata.rs
